Quantum Boltzmann Machine for Route Optimisation/pyproject.toml
[project]
name = "quantum-boltzmann-machine-for-route-optimisation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aer>=0.1.0",
    "dwave-inspector>=0.5.5",
    "dwave-neal>=0.6.0",
    "dwave-system>=1.33.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "qiskit>=2.2.2",
    "qiskit-algorithms>=0.4.0",
]
